+++ config/routes.rb
@@ -23,6 +23,7 @@ Mapserver::Application.routes.draw do
   get '/api/maps_in_point' => "api#maps_in_point"
   get '/api/select' => "api#select"
   get '/api/get_upcoming_embargoes' => "api#get_upcoming_embargoes"
+  get '/api/stats' => "api#stats"
 
   get '/filter_oris_events/:year' => "maps#filter_oris_events"