+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Rails 8.0
 - update dependencies
 - remove - annotate, turnout
+- API endpoint for stats
 
 ## [2.5.0] - 2025-05-01
 - Ruby 3.1.7